Move member commands to it's own module
